Calculator: skip null LineNumber.Value. LineCalculatedValue: accept null values #97

+++ ElectionStatistics.Tests/Core/Methods/LocationScatterplotTest.cs
@@ -81,7 +81,7 @@ namespace ElectionStatistics.Tests.Core.Methods
             {
                 Assert.Equal(new
                 {
-                    Y = calcValues[i].Value,
+                    Y = (double) calcValues[i].Value,
                     ProtocolName = protocols[i].TitleRus,
                     Id = topProtocol.Id
                 }, data[0].Values[i]);
